Move base Panel class to wagtail.admin.panels.base
